feat(select): add KolSelectWc component and update component names
- Introduced a new web component `KolSelectWc` for enhanced select functionality.
- Added `KolSelectWcTag` to component names for consistency.
- Updated localization files to simplify page entry messages in German and English.

Refs: #8653
