+++ packages/components/src/components/select/readme.md
@@ -119,7 +119,7 @@ import { xxx..., xxx..., FormatHandler, } from '@leanup/form';
 
 | Property                | Attribute        | Description                                                                                                                                                                           | Type                                                                                                                                                                                                                                                                                                                                                                                        | Default     |
 | ----------------------- | ---------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------- |
-| `_accessKey`            | `_access-key`    | Defines the key combination that can be used to trigger or focus the component’s interactive element.                                                                                 | `string \| undefined`                                                                                                                                                                                                                                                                                                                                                                       | `undefined` |
+| `_accessKey`            | `_access-key`    | Defines the key combination that can be used to trigger or focus the component's interactive element.                                                                                 | `string \| undefined`                                                                                                                                                                                                                                                                                                                                                                       | `undefined` |
 | `_alert`                | `_alert`         | <span style="color:red">**[DEPRECATED]**</span> Will be removed in v3. Use automatic behaviour instead.<br/><br/>Defines whether the screen-readers should read out the notification. | `boolean \| undefined`                                                                                                                                                                                                                                                                                                                                                                      | `undefined` |
 | `_disabled`             | `_disabled`      | Makes the element not focusable and ignore all events.                                                                                                                                | `boolean \| undefined`                                                                                                                                                                                                                                                                                                                                                                      | `false`     |
 | `_error`                | `_error`         | <span style="color:red">**[DEPRECATED]**</span> Will be removed in v3. Use `msg` instead.<br/><br/>Defines the error message text.                                                    | `string \| undefined`                                                                                                                                                                                                                                                                                                                                                                       | `undefined` |
