+++ packages/components/src/components/select/test/snapshot.spec.tsx
@@ -1,6 +1,7 @@
 import { KolSelectTag } from '../../../core/component-names';
 import type { SelectProps } from '../../../schema';
 import { executeInputSnapshotTests } from '../../../utils/testing';
+import { KolSelectWc } from '../component';
 
 import { KolSelect } from '../shadow';
 
@@ -20,11 +21,11 @@ const options = [
 	},
 ];
 
-executeInputSnapshotTests<SelectProps>(KolSelectTag, [KolSelect], {
+executeInputSnapshotTests<SelectProps>(KolSelectTag, [KolSelect, KolSelectWc], {
 	_options: options,
 });
 
-executeInputSnapshotTests<SelectProps>(KolSelectTag, [KolSelect], {
+executeInputSnapshotTests<SelectProps>(KolSelectTag, [KolSelect, KolSelectWc], {
 	_options: options,
 	_multiple: true,
 });
