+++ packages/components/src/core/component-names.ts
@@ -47,6 +47,7 @@ export let KolPopoverWcTag = 'kol-popover-wc' as const;
 export let KolProgressTag = 'kol-progress' as const;
 export let KolQuoteTag = 'kol-quote' as const;
 export let KolSelectTag = 'kol-select' as const;
+export let KolSelectWcTag = 'kol-select-wc' as const;
 export let KolSingleSelectTag = 'kol-single-select' as const;
 export let KolSkipNavTag = 'kol-skip-nav' as const;
 export let KolSpanWcTag = 'kol-span-wc' as const;
@@ -118,6 +119,7 @@ export const setCustomTagNames = (transformTagName: (tagName: string) => string)
 	KolProgressTag = transformTagName(KolProgressTag as string) as 'kol-progress';
 	KolQuoteTag = transformTagName(KolQuoteTag as string) as 'kol-quote';
 	KolSelectTag = transformTagName(KolSelectTag as string) as 'kol-select';
+	KolSelectWcTag = transformTagName(KolSelectWcTag as string) as 'kol-select-wc';
 	KolSingleSelectTag = transformTagName(KolSingleSelectTag as string) as 'kol-single-select';
 	KolSkipNavTag = transformTagName(KolSkipNavTag as string) as 'kol-skip-nav';
 	KolSpanWcTag = transformTagName(KolSpanWcTag as string) as 'kol-span-wc';
